refactor: server/index.ts からSocket.ioハンドラとTTS処理を分離
Socket.ioイベントハンドラ(user-input/disconnect)とprocessSentenceを
server/handlers/socketConnectionHandler.ts へ移動し、
server/index.ts をサーバー起動・配線のみを担う薄いエントリーポイントにする。
